+++ R/messager.R
@@ -1,6 +1,0 @@
-messager <- function(..., v = TRUE) {
-    msg <- paste(...)
-    if (v) {
-        message(msg)
-    }
-}